# backend/src/swagger.bundle.yaml
openapi: 3.0.0
info:
  title: VaiPelaSombra API
  version: 1.0.0
  description: API documentation for VaiPelaSombra Backend
servers:
  - url: http://localhost:3000
    description: Local server
paths:
...
  /itineraries/{itineraryId}/transports:
    get:
      tags:
        - Transports
      summary: Get all transports for a given itinerary
      operationId: getTransportsByItinerary
      parameters:
        - name: itineraryId
          in: path
          required: true
          description: ID of the itinerary to fetch transports for
          schema:
            type: integer
      responses:
        '200':
          description: A list of transports for the itinerary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transport'
    post:
      tags:
        - Transports
      summary: Create a new transport for an itinerary
      operationId: createTransport
      parameters:
        - name: itineraryId
          in: path
          required: true
          description: ID of the itinerary to associate the transport with
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransportInput'
      responses:
        '201':
          description: Transport created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transport'
  /transport/{id}:
    get:
      tags:
        - Transports
      summary: Get transport by ID
      operationId: getTransportById
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the transport
          schema:
            type: integer
      responses:
        '200':
          description: Transport found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transport'
        '404':
          description: Transport not found
    put:
      tags:
        - Transports
      summary: Update transport by ID
      operationId: updateTransport
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the transport to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransportInput'
      responses:
        '200':
          description: Transport updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transport'
        '404':
          description: Transport not found
    delete:
      tags:
        - Transports
      summary: Delete transport by ID
      operationId: deleteTransport
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the transport to delete
          schema:
            type: integer
      responses:
        '204':
          description: Transport deleted successfully
        '404':
          description: Transport not found
        '500':
          description: Server error
...
components:
  schemas:
...
    CreateTransportInput:
      type: object
      required:
        - type
        - cost
        - itineraryId
      properties:
        type:
          type: string
          example: Airplane
        cost:
          type: number
          format: float
          example: 1234.56
        departure:
          type: string
          format: date-time
          example: '2025-08-01T08:00:00Z'
          nullable: true
        arrival:
          type: string
          format: date-time
          example: '2025-08-01T12:00:00Z'
          nullable: true
        duration:
          type: string
          example: 4h
          nullable: true
        description:
          type: string
          example: Direct flight to destination
          nullable: true
        itineraryId:
          type: integer
    UpdateTransportInput:
      type: object
      properties:
        type:
          type: string
          example: Bus
        cost:
          type: number
          format: float
          example: 250
        departure:
          type: string
          format: date-time
          example: '2025-08-03T07:00:00Z'
          nullable: true
        arrival:
          type: string
          format: date-time
          example: '2025-08-03T13:00:00Z'
          nullable: true
        duration:
          type: string
          example: 6h
          nullable: true
        description:
          type: string
          example: Bus trip with stops
          nullable: true
        itineraryId:
          type: integer
          nullable: true
    Transport:
      type: object
      properties:
        id:
          type: integer
          example: 5
        type:
          type: string
          example: Airplane
        cost:
          type: number
          format: float
          example: 1234.56
        departure:
          type: string
          format: date-time
          example: '2025-08-01T08:00:00Z'
          nullable: true
        arrival:
          type: string
          format: date-time
          example: '2025-08-01T12:00:00Z'
          nullable: true
        duration:
          type: string
          example: 4h
          nullable: true
        description:
          type: string
          example: Direct flight to destination
          nullable: true
        itineraryId:
          type: integer
          example: 1